+++ packages/fxa-shared/db/models/auth/base-auth.ts
@@ -64,7 +64,6 @@ export enum Proc {
   UpsertAvailableCommands = 'upsertAvailableCommand_1',
   VerifyEmail = 'verifyEmail_9',
   VerifyToken = 'verifyToken_3',
-  VerifyTokenCode = 'verifyTokenCode_2',
   VerifyTokenWithMethod = 'verifyTokensWithMethod_3',
 }