+++ packages/fxa-shared/db/models/auth/index.ts
@@ -22,6 +22,7 @@ import { EmailType } from './email-type';
 import { SentEmail } from './sent-email';
 import { SecurityEvent } from './security-event';
 import { PruneTokens } from './prune-token';
+import { RelyingParty } from './relying-party';
 
 export type PayPalBillingAgreementStatusType =
   | 'Pending'
@@ -252,4 +253,5 @@ export {
   SentEmail,
   SessionToken,
   TotpToken,
+  RelyingParty,
 };