+++ packages/fxa-geodb/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change history
 
+## 1.234.1
+
+No changes.
+
+## 1.234.0
+
+No changes.
+
 ## 1.233.1
 
 No changes.